Homework 1: 1. Select list of user's git repos into file 2. Use any API with authentification to get the information into file
